Adjusted model, added draft version of exemplary building
Project still contains errors

+++ lib/models/location.dart
@@ -1,7 +1,5 @@
-import 'dart:ffi';
-
-class Location{
-  Float lat;
-  Float lon;
+class Location {
+  double lat;
+  double lon;
   Location({this.lat, this.lon});
 }
